pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "archetype-core-etl"
version = "0.1.0"
description = "Production-grade ETL platform orchestrated by Airflow, powered by Databricks, provisioned by Terraform."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Archetype Data Platform" },
]
keywords = ["etl", "airflow", "databricks", "data-engineering", "pipelines"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
]
dependencies = [
    "pydantic~=2.12",
    "pydantic-settings~=2.13",
    "python-dotenv~=1.0",
    "python-json-logger~=4.1",
    "boto3~=1.42",
    "great-expectations~=1.16",
    "pandas~=2.0",
    "databricks-sdk~=0.102",
    "psycopg2-binary~=2.9",
    "faker~=40.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow~=3.2",
    "apache-airflow-providers-celery~=4.0",
    "apache-airflow-providers-fab~=1.0",
]
databricks = [
    # databricks-connect added when Databricks Connect support is needed.
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock~=3.14",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "boto3-stubs[s3,kinesis,bedrock-runtime]~=1.42",
    "pandas-stubs~=2.0",
    "types-psycopg2~=2.9",
]

[project.urls]
Homepage = "https://github.com/marlungu/archetype-core-etl"
Repository = "https://github.com/marlungu/archetype-core-etl"
Issues = "https://github.com/marlungu/archetype-core-etl/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/archetype_core_etl"]

[tool.hatch.build.targets.sdist]
include = [
    "src/archetype_core_etl",
    "README.md",
    "pyproject.toml",
]

# ---------- Ruff ----------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "C4",  # flake8-comprehensions
    "PL",  # pylint
    "RUF", # ruff-specific
    "S",   # flake8-bandit security checks
]
ignore = [
    "E501",   # handled by formatter
    "PLR0913", # allow >5 args in constructors with keyword-only params
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004", "S101"]
"dags/**/*.py" = ["PLC0415", "PLR0915"]  # deferred imports and long DAG functions are intentional
"src/archetype_core_etl/load/delta_writer.py" = ["S608"]  # table_fqn is internal, values are parameterized
"src/archetype_core_etl/config/settings.py" = ["PLW0108"]  # lambdas required for mypy forward-ref resolution

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------- Mypy ----------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ---------- Pytest ----------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=archetype_core_etl",
    "--cov-report=term-missing",
]
markers = [
    "unit: fast, isolated unit tests",
    "integration: tests that hit external systems",
    "slow: tests that take more than a second",
]

# ---------- Coverage ----------
[tool.coverage.run]
branch = true
source = ["src/archetype_core_etl"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
